+++ exporters/metric/cortex/cortex.go
@@ -23,7 +23,6 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/gogo/protobuf/proto"
 	"github.com/golang/snappy"
 	"github.com/prometheus/prometheus/prompb"
 
@@ -34,15 +33,23 @@ import (
 	"go.opentelemetry.io/otel/sdk/export/metric"
 	export "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
+	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric/aggregator/histogram"
 	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
 	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
 	"go.opentelemetry.io/otel/sdk/metric/selector/simple"
+	"go.opentelemetry.io/otel/sdk/resource"
 )
 
 // Exporter forwards metrics to a Cortex instance
 type Exporter struct {
 	config Config
+}
+
+type exportData struct {
+	export.Record
+
+	Resource *resource.Resource
 }
 
 // ExportKindFor returns CumulativeExporter so the Processor correctly aggregates data
@@ -51,8 +58,8 @@ func (e *Exporter) ExportKindFor(*apimetric.Descriptor, aggregation.Kind) metric
 }
 
 // Export forwards metrics to Cortex from the SDK
-func (e *Exporter) Export(_ context.Context, checkpointSet metric.CheckpointSet) error {
-	timeseries, err := e.ConvertToTimeSeries(checkpointSet)
+func (e *Exporter) Export(_ context.Context, res *resource.Resource, checkpointSet metric.InstrumentationLibraryReader) error {
+	timeseries, err := e.ConvertToTimeSeries(res, checkpointSet)
 	if err != nil {
 		return err
 	}
@@ -95,8 +102,8 @@ func NewExportPipeline(config Config, options ...controller.Option) (*controller
 		return nil, err
 	}
 
-	pusher := controller.New(
-		processor.New(
+	cont := controller.New(
+		processor.NewFactory(
 			simple.NewWithHistogramDistribution(
 				histogram.WithExplicitBoundaries(config.HistogramBoundaries),
 			),
@@ -105,68 +112,74 @@ func NewExportPipeline(config Config, options ...controller.Option) (*controller
 		append(options, controller.WithExporter(exporter))...,
 	)
 
-	return pusher, pusher.Start(context.TODO())
+	return cont, cont.Start(context.TODO())
 }
 
 // InstallNewPipeline registers a push Controller's MeterProvider globally.
 func InstallNewPipeline(config Config, options ...controller.Option) (*controller.Controller, error) {
-	pusher, err := NewExportPipeline(config, options...)
+	cont, err := NewExportPipeline(config, options...)
 	if err != nil {
 		return nil, err
 	}
-	global.SetMeterProvider(pusher.MeterProvider())
-	return pusher, nil
+	global.SetMeterProvider(cont)
+	return cont, nil
 }
 
-// ConvertToTimeSeries converts a CheckpointSet to a slice of TimeSeries pointers
+// ConvertToTimeSeries converts a InstrumentationLibraryReader to a slice of TimeSeries pointers
 // Based on the aggregation type, ConvertToTimeSeries will call helper functions like
 // convertFromSum to generate the correct number of TimeSeries.
-func (e *Exporter) ConvertToTimeSeries(checkpointSet export.CheckpointSet) ([]*prompb.TimeSeries, error) {
+func (e *Exporter) ConvertToTimeSeries(res *resource.Resource, checkpointSet export.InstrumentationLibraryReader) ([]prompb.TimeSeries, error) {
 	var aggError error
-	var timeSeries []*prompb.TimeSeries
+	var timeSeries []prompb.TimeSeries
 
 	// Iterate over each record in the checkpoint set and convert to TimeSeries
-	aggError = checkpointSet.ForEach(e, func(record metric.Record) error {
-		// Convert based on aggregation type
-		agg := record.Aggregation()
+	aggError = checkpointSet.ForEach(func(library instrumentation.Library, reader export.Reader) error {
+		return reader.ForEach(e, func(record metric.Record) error {
+			// Convert based on aggregation type
+			edata := exportData{
+				Resource: res,
+				Record:   record,
+			}
+			agg := record.Aggregation()
 
-		// The following section uses loose type checking to determine how to
-		// convert aggregations to timeseries. More "expensive" timeseries are
-		// checked first.
-		//
-		// See the Aggregator Kind for more information
-		// https://github.com/open-telemetry/opentelemetry-go/blob/main/sdk/export/metric/aggregation/aggregation.go#L123-L138
-		if histogram, ok := agg.(aggregation.Histogram); ok {
-			tSeries, err := convertFromHistogram(record, histogram)
-			if err != nil {
-				return err
-			}
-			timeSeries = append(timeSeries, tSeries...)
-		} else if sum, ok := agg.(aggregation.Sum); ok {
-			tSeries, err := convertFromSum(record, sum)
-			if err != nil {
-				return err
-			}
-			timeSeries = append(timeSeries, tSeries)
-			if minMaxSumCount, ok := agg.(aggregation.MinMaxSumCount); ok {
-				tSeries, err := convertFromMinMaxSumCount(record, minMaxSumCount)
+			// The following section uses loose type checking to determine how to
+			// convert aggregations to timeseries. More "expensive" timeseries are
+			// checked first.
+			//
+			// See the Aggregator Kind for more information
+			// https://github.com/open-telemetry/opentelemetry-go/blob/main/sdk/export/metric/aggregation/aggregation.go#L123-L138
+			if histogram, ok := agg.(aggregation.Histogram); ok {
+				tSeries, err := convertFromHistogram(edata, histogram)
 				if err != nil {
 					return err
 				}
 				timeSeries = append(timeSeries, tSeries...)
+			} else if sum, ok := agg.(aggregation.Sum); ok {
+				tSeries, err := convertFromSum(edata, sum)
+				if err != nil {
+					return err
+				}
+				timeSeries = append(timeSeries, tSeries)
+				if minMaxSumCount, ok := agg.(aggregation.MinMaxSumCount); ok {
+					tSeries, err := convertFromMinMaxSumCount(edata, minMaxSumCount)
+					if err != nil {
+						return err
+					}
+					timeSeries = append(timeSeries, tSeries...)
+				}
+			} else if lastValue, ok := agg.(aggregation.LastValue); ok {
+				tSeries, err := convertFromLastValue(edata, lastValue)
+				if err != nil {
+					return err
+				}
+				timeSeries = append(timeSeries, tSeries)
+			} else {
+				// Report to the user when no conversion was found
+				fmt.Printf("No conversion found for record: %s\n", edata.Descriptor().Name())
 			}
-		} else if lastValue, ok := agg.(aggregation.LastValue); ok {
-			tSeries, err := convertFromLastValue(record, lastValue)
-			if err != nil {
-				return err
-			}
-			timeSeries = append(timeSeries, tSeries)
-		} else {
-			// Report to the user when no conversion was found
-			fmt.Printf("No conversion found for record: %s\n", record.Descriptor().Name())
-		}
 
-		return nil
+			return nil
+		})
 	})
 
 	// Check if error was returned in checkpointSet.ForEach()
@@ -178,83 +191,83 @@ func (e *Exporter) ConvertToTimeSeries(checkpointSet export.CheckpointSet) ([]*p
 }
 
 // createTimeSeries is a helper function to create a timeseries from a value and attributes
-func createTimeSeries(record metric.Record, value number.Number, valueNumberKind number.Kind, extraAttributes ...attribute.KeyValue) *prompb.TimeSeries {
+func createTimeSeries(edata exportData, value number.Number, valueNumberKind number.Kind, extraAttributes ...attribute.KeyValue) prompb.TimeSeries {
 	sample := prompb.Sample{
 		Value:     value.CoerceToFloat64(valueNumberKind),
-		Timestamp: int64(time.Nanosecond) * record.EndTime().UnixNano() / int64(time.Millisecond),
+		Timestamp: int64(time.Nanosecond) * edata.EndTime().UnixNano() / int64(time.Millisecond),
 	}
 
-	attributes := createLabelSet(record, extraAttributes...)
+	attributes := createLabelSet(edata, extraAttributes...)
 
-	return &prompb.TimeSeries{
+	return prompb.TimeSeries{
 		Samples: []prompb.Sample{sample},
 		Labels:  attributes,
 	}
 }
 
 // convertFromSum returns a single TimeSeries based on a Record with a Sum aggregation
-func convertFromSum(record metric.Record, sum aggregation.Sum) (*prompb.TimeSeries, error) {
+func convertFromSum(edata exportData, sum aggregation.Sum) (prompb.TimeSeries, error) {
 	// Get Sum value
 	value, err := sum.Sum()
 	if err != nil {
-		return nil, err
+		return prompb.TimeSeries{}, err
 	}
 
 	// Create TimeSeries. Note that Cortex requires the name attribute to be in the format
 	// "__name__". This is the case for all time series created by this exporter.
-	name := sanitize(record.Descriptor().Name())
-	numberKind := record.Descriptor().NumberKind()
-	tSeries := createTimeSeries(record, value, numberKind, attribute.String("__name__", name))
+	name := sanitize(edata.Descriptor().Name())
+	numberKind := edata.Descriptor().NumberKind()
+	tSeries := createTimeSeries(edata, value, numberKind, attribute.String("__name__", name))
 
 	return tSeries, nil
 }
 
 // convertFromLastValue returns a single TimeSeries based on a Record with a LastValue aggregation
-func convertFromLastValue(record metric.Record, lastValue aggregation.LastValue) (*prompb.TimeSeries, error) {
+func convertFromLastValue(edata exportData, lastValue aggregation.LastValue) (prompb.TimeSeries, error) {
 	// Get value
 	value, _, err := lastValue.LastValue()
 	if err != nil {
-		return nil, err
+		return prompb.TimeSeries{}, err
 	}
 
 	// Create TimeSeries
-	name := sanitize(record.Descriptor().Name())
-	numberKind := record.Descriptor().NumberKind()
-	tSeries := createTimeSeries(record, value, numberKind, attribute.String("__name__", name))
+	name := sanitize(edata.Descriptor().Name())
+	numberKind := edata.Descriptor().NumberKind()
+	tSeries := createTimeSeries(edata, value, numberKind, attribute.String("__name__", name))
 
 	return tSeries, nil
 }
 
 // convertFromMinMaxSumCount returns 4 TimeSeries for the min, max, sum, and count from the mmsc aggregation
-func convertFromMinMaxSumCount(record metric.Record, minMaxSumCount aggregation.MinMaxSumCount) ([]*prompb.TimeSeries, error) {
-	numberKind := record.Descriptor().NumberKind()
+func convertFromMinMaxSumCount(edata exportData, minMaxSumCount aggregation.MinMaxSumCount) ([]prompb.TimeSeries, error) {
+	numberKind := edata.Descriptor().NumberKind()
 
 	// Convert Min
 	min, err := minMaxSumCount.Min()
 	if err != nil {
 		return nil, err
 	}
-	name := sanitize(record.Descriptor().Name() + "_min")
-	minTimeSeries := createTimeSeries(record, min, numberKind, attribute.String("__name__", name))
+	name := sanitize(edata.Descriptor().Name() + "_min")
+	minTimeSeries := createTimeSeries(edata, min, numberKind, attribute.String("__name__", name))
 
 	// Convert Max
 	max, err := minMaxSumCount.Max()
 	if err != nil {
 		return nil, err
 	}
-	name = sanitize(record.Descriptor().Name() + "_max")
-	maxTimeSeries := createTimeSeries(record, max, numberKind, attribute.String("__name__", name))
+	name = sanitize(edata.Descriptor().Name() + "_max")
+	maxTimeSeries := createTimeSeries(edata, max, numberKind, attribute.String("__name__", name))
 
 	// Convert Count
 	count, err := minMaxSumCount.Count()
 	if err != nil {
 		return nil, err
 	}
-	name = sanitize(record.Descriptor().Name() + "_count")
-	countTimeSeries := createTimeSeries(record, number.NewInt64Number(int64(count)), number.Int64Kind, attribute.String("__name__", name))
+	name = sanitize(edata.Descriptor().Name() + "_count")
+	countTimeSeries := createTimeSeries(edata, number.NewInt64Number(int64(count)), number.Int64Kind, attribute.String("__name__", name))
 
 	// Return all timeSeries
-	tSeries := []*prompb.TimeSeries{
+	tSeries := []prompb.TimeSeries{
 		minTimeSeries, maxTimeSeries, countTimeSeries,
 	}
 
@@ -262,17 +275,17 @@ func convertFromMinMaxSumCount(record metric.Record, minMaxSumCount aggregation.
 }
 
 // convertFromHistogram returns len(histogram.Buckets) timeseries for a histogram aggregation
-func convertFromHistogram(record metric.Record, histogram aggregation.Histogram) ([]*prompb.TimeSeries, error) {
-	var timeSeries []*prompb.TimeSeries
-	metricName := sanitize(record.Descriptor().Name())
-	numberKind := record.Descriptor().NumberKind()
+func convertFromHistogram(edata exportData, histogram aggregation.Histogram) ([]prompb.TimeSeries, error) {
+	var timeSeries []prompb.TimeSeries
+	metricName := sanitize(edata.Descriptor().Name())
+	numberKind := edata.Descriptor().NumberKind()
 
 	// Create Sum TimeSeries
 	sum, err := histogram.Sum()
 	if err != nil {
 		return nil, err
 	}
-	sumTimeSeries := createTimeSeries(record, sum, numberKind, attribute.String("__name__", metricName+"_sum"))
+	sumTimeSeries := createTimeSeries(edata, sum, numberKind, attribute.String("__name__", metricName+"_sum"))
 	timeSeries = append(timeSeries, sumTimeSeries)
 
 	// Handle Histogram buckets
@@ -294,7 +307,7 @@ func convertFromHistogram(record metric.Record, histogram aggregation.Histogram)
 		boundaryStr := strconv.FormatFloat(boundary, 'f', -1, 64)
 
 		// Create timeSeries and append
-		boundaryTimeSeries := createTimeSeries(record, number.NewFloat64Number(totalCount), number.Float64Kind, attribute.String("__name__", metricName), attribute.String("le", boundaryStr))
+		boundaryTimeSeries := createTimeSeries(edata, number.NewFloat64Number(totalCount), number.Float64Kind, attribute.String("__name__", metricName), attribute.String("le", boundaryStr))
 		timeSeries = append(timeSeries, boundaryTimeSeries)
 	}
 
@@ -304,9 +317,9 @@ func convertFromHistogram(record metric.Record, histogram aggregation.Histogram)
 	// Create a timeSeries for the +inf bucket and total count
 	// These are the same and are both required by Prometheus-based backends
 
-	upperBoundTimeSeries := createTimeSeries(record, number.NewFloat64Number(totalCount), number.Float64Kind, attribute.String("__name__", metricName), attribute.String("le", "+inf"))
+	upperBoundTimeSeries := createTimeSeries(edata, number.NewFloat64Number(totalCount), number.Float64Kind, attribute.String("__name__", metricName), attribute.String("le", "+inf"))
 
-	countTimeSeries := createTimeSeries(record, number.NewFloat64Number(totalCount), number.Float64Kind, attribute.String("__name__", metricName+"_count"))
+	countTimeSeries := createTimeSeries(edata, number.NewFloat64Number(totalCount), number.Float64Kind, attribute.String("__name__", metricName+"_count"))
 
 	timeSeries = append(timeSeries, upperBoundTimeSeries)
 	timeSeries = append(timeSeries, countTimeSeries)
@@ -316,13 +329,13 @@ func convertFromHistogram(record metric.Record, histogram aggregation.Histogram)
 
 // createLabelSet combines attributes from a Record, resource, and extra attributes to create a
 // slice of prompb.Label.
-func createLabelSet(record metric.Record, extraAttributes ...attribute.KeyValue) []*prompb.Label {
+func createLabelSet(edata exportData, extraAttributes ...attribute.KeyValue) []prompb.Label {
 	// Map ensure no duplicate label names.
 	labelMap := map[string]prompb.Label{}
 
 	// mergeAttributes merges Record and Resource attributes into a single set, giving precedence
 	// to the record's attributes.
-	mi := attribute.NewMergeIterator(record.Labels(), record.Resource().Set())
+	mi := attribute.NewMergeIterator(edata.Labels(), edata.Resource.Set())
 	for mi.Next() {
 		attribute := mi.Label()
 		key := string(attribute.Key)
@@ -350,10 +363,9 @@ func createLabelSet(record metric.Record, extraAttributes ...attribute.KeyValue)
 	}
 
 	// Create slice of labels from labelMap and return
-	res := make([]*prompb.Label, 0, len(labelMap))
+	res := make([]prompb.Label, 0, len(labelMap))
 	for _, lb := range labelMap {
-		currentLabel := lb
-		res = append(res, &currentLabel)
+		res = append(res, lb)
 	}
 
 	return res
@@ -387,17 +399,19 @@ func (e *Exporter) addHeaders(req *http.Request) error {
 }
 
 // buildMessage creates a Snappy-compressed protobuf message from a slice of TimeSeries.
-func (e *Exporter) buildMessage(timeseries []*prompb.TimeSeries) ([]byte, error) {
+func (e *Exporter) buildMessage(timeseries []prompb.TimeSeries) ([]byte, error) {
 	// Wrap the TimeSeries as a WriteRequest since Cortex requires it.
 	writeRequest := &prompb.WriteRequest{
 		Timeseries: timeseries,
 	}
 
 	// Convert the struct to a slice of bytes and then compress it.
-	message, err := proto.Marshal(writeRequest)
+	message := make([]byte, writeRequest.Size())
+	written, err := writeRequest.MarshalToSizedBuffer(message)
 	if err != nil {
 		return nil, err
 	}
+	message = message[:written]
 	compressed := snappy.Encode(nil, message)
 
 	return compressed, nil